+++ docs/changelog.rst
@@ -26,6 +26,7 @@ In progress (unreleased)
 * :gh:issue:`1423` tests: Group and unify naming of connection benchmarks
 * :gh:issue:`1424` tests: Parameterize connection benchmarks
 * :gh:issue:`1424` tests: Standardise output of connection benchmarks
+* :gh:issue:`1424` tests: Parameterize throughput benchmark
 
 
 v0.3.37 (2026-01-08)